Add a codemod for dataclass changes in 3.11
Python 3.11 changed the field default mutability check for dataclasses to only allow defaults which are hashable.

This codemod helps with the migration by changing all default values that aren't obviously hashable to use `default_factory` instead.

Note: it's impossible to accurately determine if a particular expression produces a hashable value in a codemod, so the codemod significantly over-approximates what's unhashable.